Adds branch based lookup table for virtual tags
During development it might be desirable to be able to already
designate a commit that is the start of a future release version
and generating version strings during development relative to
that start commit, possibly also with a more development friendly
versioning scheme than the regular one used in a project.

This patch adds this possibility by introducing a lookup file
that allows specifying virtual tags plus reference commits
on a branch basis. Optionally it is also possible to override
the render scheme configured in versioneer's setup.cfg with a
branch-specific one.

The lookup file is processed line by line, the first line that
matches the current branch wins. By only specifying a branch
but no further parameters it is possible to "exit early" and
fall back to regular versioneer handling. It is also possible
to only override the rendering scheme without specifying a
virtual tag.

The lookup file to use can be specified via a new parameter
in versioneer's config. By default a file ".versioneer-lookup"
in the project's root will be used.

An example would be the following file:

    master
    next                          pep440-pre
    maintenance  1.1  838ba8c83
    devel        1.2  0204a90c0   pep440-dev

This would have the following effects depending on the
selected branch:

  * the "master" branch would use regular versioning as
    provided by stock versioneer
  * the "next" branch would use regular versioning but
    render using the pep440-pre render style
  * the "maintenance" branch would behave as if the provided
    commit hash 838ba8c83 was tagged 1.1 and use stock
    render style
  * the "devel" branch would behave as if the provided
    commit hash 9294a90c0 was tagged 1.2 and use the
    pep440-dev style
  * all other branches would just use regular versioning
    as provided by the stock versioneer

The lookup file also allows comments, started with the
hash (#) character and ended by a new line