New style pep440-tag
A bit like pep440, but (usually) without any bits other than the tag.

TAG

Falls back to this if no tag could be found:

0.postDISTANCE[.dev0]+gHEX

and to this if the working copy is dirty:

TAG.postDISTANCE.dev0+gHEX

Doesn't sound overly useful at first glance (why not just spec the
version directly in the source if we don't want it to adjust
automatically) but in projects that to per-branch version schemes this
can be of huge value if the master branch's version should not change
on ever tiny commit done for repository maintenance reasons. I'm
looking at you, Github - I want to be able to change my CONTRIBUTING.md,
the issue template and such without either waiting for the next stable
release or automatically changing the stable version displayed to end
users along side it ;)